+++ package.json
@@ -42,7 +42,8 @@
     "clsx": "^2.1.1",
     "cmdk": "^1.0.0",
     "date-fns": "^3.6.0",
-    "embla-carousel-react": "^8.3.1",
+    "embla-carousel-autoplay": "^8.5.2",
+    "embla-carousel-react": "^8.5.2",
     "libphonenumber-js": "^1.11.17",
     "lucide-react": "^0.446.0",
     "next": "15.0.3",
